refactor(BridgeGame): 폴더 경로 변경
BridgeGame.js는 파일의 경로를 처음 주어진 것에서 변경하면 안된다.
따라서 원래 위치로 복귀시키고, 이 파일을 참조하는 곳에서 경로를 수정한다.
