+++ src/BridgeMaker.js
@@ -1,4 +1,4 @@
-const { GAME } = require('../constant/Constants');
+const { GAME } = require('./constant/Constants');
 
 /**
  * 다리의 길이를 입력 받아서 다리를 생성해주는 역할을 한다.
